Probably: Fixes broken notifs (#66)
* Theoretically fixes broken notifications

I'm not very good at rust and this code is a little hard to read, so I'm not sure if I did all of this right - would appreciate a code review. Basically, slack allows you to pass a value to the `text` parameter *after* you have already passed a value to `blocks`, and it will send that as what it calls "fallback text". Slack will use this in places they can't display blocks, like in notifications or the mentions tab of the mobile app.

I changed some indenting on this to conform with existing formatting - I recommend appending `?w=1` to the GitHub URL so that it shows non-whitespace changes.

* update special user message for grammar

* fix for fix for special user message formatting

* run cargo fmt on code

* run cargo fmt on hacksteader.rs

* fix leveling up and farm cycle notifications

* fix blank strings which could potentially cause missing notifications

* fix purchase message + formatting