fix: Sort apply item modal so no items are left out (#113)
* Paginate recipes (#106)

* Use all seeds in the seed picking GUI (#90)

* Fix new hcor yield API

* feat: Yields now produce xp

* fix: crafting with multiple outputs only providing one

* Make flavor text a little less exuberant (#91)

* Update version to 1.0.7

* Update version to 1.0.8

* fix: Clarify market message

* Theoretically fixes broken notifications

I'm not very good at rust and this code is a little hard to read, so I'm not sure if I did all of this right - would appreciate a code review. Basically, slack allows you to pass a value to the `text` parameter *after* you have already passed a value to `blocks`, and it will send that as what it calls "fallback text". Slack will use this in places they can't display blocks, like in notifications or the mentions tab of the mobile app.

I changed some indenting on this to conform with existing formatting - I recommend appending `?w=1` to the GitHub URL so that it shows non-whitespace changes.

* update special user message for grammar

* fix for fix for special user message formatting

* run cargo fmt on code

* run cargo fmt on hacksteader.rs

* fix leveling up and farm cycle notifications

* fix blank strings which could potentially cause missing notifications

* fix purchase message + formatting

* Fix notification messages to be more clear to actual purpose

* update notifications for grammar and proper UID parsing

* fix unclear market message

* Fix grammar in response

* bump hcor version number to 1.0.3

* Implement pagination for the recipe GUI

* Save the locked recipes for the next page, always

* fmt

* fix tests

Co-authored-by: Wei Gao <gaowmhs1@gmail.com>
Co-authored-by: Owen Salter <owen@devosmium.xyz>
Co-authored-by: Luke Carapezza <me@lukec.me>

* fix: Sort apply item modal so no items are left out

Co-authored-by: Wei Gao <gaowmhs1@gmail.com>
Co-authored-by: Owen Salter <owen@devosmium.xyz>
Co-authored-by: Luke Carapezza <me@lukec.me>